third_party/blink/web_tests/platform/generic/fast/css/invalidation/scrollbar-pseudo-expected.txt

Style invalidation for scrollbar pseudo elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(scroller1, '::-webkit-scrollbar').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(scroller1, '::-webkit-scrollbar').backgroundColor is green
PASS getComputedStyle(scroller2, '::-webkit-scrollbar').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(scroller2, '::-webkit-scrollbar').backgroundColor is green
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-button').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-button').backgroundColor is green
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-button').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-button').backgroundColor is green
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-corner').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-corner').backgroundColor is green
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-corner').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-corner').backgroundColor is green
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-thumb').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-thumb').backgroundColor is green
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-thumb').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-thumb').backgroundColor is green
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-track').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-track').backgroundColor is green
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-track').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-track').backgroundColor is green
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-track-piece').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 4
PASS getComputedStyle(scroller1, '::-webkit-scrollbar-track-piece').backgroundColor is green
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-track-piece').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 4
PASS getComputedStyle(scroller2, '::-webkit-scrollbar-track-piece').backgroundColor is green
PASS getComputedStyle(scroller1, '::-webkit-resizer').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(scroller1, '::-webkit-resizer').backgroundColor is green
PASS getComputedStyle(scroller2, '::-webkit-resizer').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(scroller2, '::-webkit-resizer').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
